split the project into modules
