+++ js/data.js
@@ -1,0 +1,48 @@
+const NAMES = [
+  'Александр', 'Мария', 'Дмитрий', 'Екатерина', 'Иван',
+  'Анна', 'Максим', 'Ольга', 'Сергей', 'Татьяна',
+  'Анастасия', 'Павел', 'Светлана', 'Николай', 'Елена',
+  'Артем', 'Юлия', 'Роман', 'Дарья', 'Алексей',
+  'Ксения', 'Виктор', 'Олег', 'Ирина', 'Валентина'
+];
+
+let messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.Как можно было поймать такой неудачный момент?!'
+];
+
+let photoDescriptions = [
+  'Солнечный закат на пляже.',
+  'Горы, покрытые снежным покровом.',
+  'Яркие цветы весной.',
+  'Ночной город с огнями.',
+  'Спокойное озеро в лесу.',
+  'Семейный пикник на природе.',
+  'Друзья, наслаждающиеся вечеринкой.',
+  'Уютный уголок в кафе.',
+  'Старинная архитектура города.',
+  'Путешествие по живописным дорогам.',
+  'Улыбающиеся дети на прогулке.',
+  'Красивая радуга после дождя.',
+  'Пейзаж с полями и деревьями.',
+  'Зимний лес с падающим снегом.',
+  'Кулинарные шедевры на столе.',
+  'Спортивное событие с болельщиками.',
+  'Морская прогулка на яхте.',
+  'Фотография любимого питомца.',
+  'Творческий процесс художника.',
+  'Светлое утро с чашкой кофе.',
+  'Уютный вечер с книгой.',
+  'Забавные моменты с друзьями.',
+  'Классическая музыка на концерте.',
+  'Дорога, ведущая в никуда.',
+  'Путешествие по историческим местам.',
+  'Счастливые моменты на свадьбе.'
+];
+
+export {NAMES, messages, photoDescriptions};
