+++ package.json
@@ -23,5 +23,6 @@
   "engines": {
     "node": "^20",
     "npm": "^10"
-  }
+  },
+  "type": "module"
 }
